+++ socialhome/static/js/content.js
@@ -56,7 +56,8 @@
                 '</div>' +
             '</div>' +
             '<div class="content-actions share-action hidden" data-content-id="<%= content.id %>">' +
-                '<button class="btn btn-secondary" aria-label="' + gettext("Share") + '" type="button">' + gettext("Share") + '</button>' +
+                '<button class="btn btn-secondary <% if (!content.has_shared) { %>hidden<% } %>" data-action="unshare" aria-label="' + gettext("Unshare") + '" type="button">' + gettext("Unshare") + '</button>' +
+                '<button class="btn btn-secondary <% if (content.has_shared) { %>hidden<% } %>" data-action="share" aria-label="' + gettext("Share") + '" type="button">' + gettext("Share") + '</button>' +
             '</div>' +
             '<div class="replies-container" data-content-id="<%= content.id %>"></div>' +
             '<% if (content.is_authenticated && content.content_type === "content") { %>' +